% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoacm.R
\name{autoacm}
\alias{autoacm}
\title{Interface d'assistance à l'ACM}
\usage{
autoacm()
}
\value{
la fonction renvoie une fenetre Shiny
}
\description{
Une fonction pour faire apparaître une fenetre Shiny permettant l'import, la selection de population, la selection de variables actives et supplementaires, une exploration sommaire de l'ACM, une aide à la classification, et une passerelle vers le package explor() de Julien Barnier.
}
